+++ source/test.ts
@@ -1,10 +1,12 @@
 import { equal, deepEqual } from 'assert-helpers'
 import kava from 'kava'
 import {
+	Format,
 	secondsInHour,
 	secondsInMinute,
 	extractTimestamp,
 	replaceTimestamps,
+	stringifyTimestamp,
 } from './index.js'
 
 const extractTests = [
@@ -171,10 +173,153 @@ const extractTests = [
 	},
 ]
 
+const formatTests = [
+	{
+		input: null,
+		formats: {
+			[Format.Numeric]: null,
+			[Format.Seconds]: null,
+			[Format.Tiny]: null,
+			[Format.Short]: null,
+			[Format.Medium]: null,
+			[Format.Long]: null,
+		},
+	},
+	{
+		input: {
+			total: 0,
+		},
+		formats: {
+			[Format.Numeric]: null,
+			[Format.Seconds]: null,
+			[Format.Tiny]: null,
+			[Format.Short]: null,
+			[Format.Medium]: null,
+			[Format.Long]: null,
+		},
+	},
+	{
+		input: {
+			total: 0,
+			seconds: 0,
+		},
+		formats: {
+			[Format.Numeric]: '0:00',
+			[Format.Seconds]: '0s',
+			[Format.Tiny]: '0s',
+			[Format.Short]: '00s',
+			[Format.Medium]: '0 secs',
+			[Format.Long]: '0 seconds',
+		},
+	},
+	{
+		input: {
+			total: 0,
+			minutes: 0,
+		},
+		formats: {
+			[Format.Numeric]: '0:00',
+			[Format.Seconds]: '0s',
+			[Format.Tiny]: '0s',
+			[Format.Short]: '00s',
+			[Format.Medium]: '0 secs',
+			[Format.Long]: '0 seconds',
+		},
+	},
+	{
+		input: {
+			total: 0,
+			hours: 0,
+		},
+		formats: {
+			[Format.Numeric]: '0:00',
+			[Format.Seconds]: '0s',
+			[Format.Tiny]: '0s',
+			[Format.Short]: '00s',
+			[Format.Medium]: '0 secs',
+			[Format.Long]: '0 seconds',
+		},
+	},
+	{
+		input: {
+			total: 1,
+			seconds: 1,
+		},
+		formats: {
+			[Format.Numeric]: '0:01',
+			[Format.Seconds]: '1s',
+			[Format.Tiny]: '1s',
+			[Format.Short]: '1s',
+			[Format.Medium]: '1 secs',
+			[Format.Long]: '1 seconds',
+		},
+	},
+	{
+		input: {
+			total: 1 + secondsInMinute,
+			seconds: 1,
+			minutes: 1,
+		},
+		formats: {
+			[Format.Numeric]: '1:01',
+			[Format.Seconds]: '61s',
+			[Format.Tiny]: '1m1s',
+			[Format.Short]: '1m 1s',
+			[Format.Medium]: '1 mins 1 secs',
+			[Format.Long]: '1 minutes 1 seconds',
+		},
+	},
+	{
+		input: {
+			total: 1 + secondsInMinute + secondsInHour,
+			seconds: 1,
+			minutes: 1,
+			hours: 1,
+		},
+		formats: {
+			[Format.Numeric]: '1:01:01',
+			[Format.Seconds]: '3661s',
+			[Format.Tiny]: '1h1m1s',
+			[Format.Short]: '1h 1m 1s',
+			[Format.Medium]: '1 hours 1 mins 1 secs',
+			[Format.Long]: '1 hours 1 minutes 1 seconds',
+		},
+	},
+	{
+		input: {
+			total: secondsInHour,
+			hours: 1,
+		},
+		formats: {
+			[Format.Numeric]: '1:00:00',
+			[Format.Seconds]: '3600s',
+			[Format.Tiny]: '1h',
+			[Format.Short]: '1h',
+			[Format.Medium]: '1 hours',
+			[Format.Long]: '1 hours',
+		},
+	},
+	{
+		input: {
+			total: secondsInMinute,
+			minutes: 1,
+		},
+		formats: {
+			[Format.Numeric]: '1:00',
+			[Format.Seconds]: '60s',
+			[Format.Tiny]: '1m',
+			[Format.Short]: '1m',
+			[Format.Medium]: '1 mins',
+			[Format.Long]: '1 minutes',
+		},
+	},
+]
+
 const replaceTest = {
 	input: `
 00 01 0:0 1:0 1:2
 0:0:0 1:0:0 1:2:3
+0:00
 00s
 01s
 00m 00s
@@ -185,9 +330,10 @@ const replaceTest = {
 01 hours 02 minutes 03 secs
 01 hours 02 minutes
 02 mins`,
-	result: `
+	replaceResult: `
 00 01 {"total":0,"minutes":0,"seconds":0} {"total":60,"minutes":1,"seconds":0} {"total":62,"minutes":1,"seconds":2}
 {"total":0,"hours":0,"minutes":0,"seconds":0} {"total":3600,"hours":1,"minutes":0,"seconds":0} {"total":3723,"hours":1,"minutes":2,"seconds":3}
+{"total":0,"minutes":0,"seconds":0}
 {"total":0,"seconds":0}
 {"total":1,"seconds":1}
 {"total":0,"minutes":0,"seconds":0}
@@ -198,6 +344,20 @@ const replaceTest = {
 {"total":3723,"hours":1,"minutes":2,"seconds":3}
 {"total":3720,"hours":1,"minutes":2}
 {"total":120,"minutes":2}`,
+	stringifyResult: `
+00 01 [00s] [1m] [1m 2s]
+[00s] [1h] [1h 2m 3s]
+[00s]
+[00s]
+[1s]
+[00s]
+[00s]
+[00s]
+[1m 1s]
+[00s]
+[1h 2m 3s]
+[1h 2m]
+[2m]`,
 }
 
 kava.suite('extract-timestamp', function (suite, test) {
@@ -212,12 +372,36 @@ kava.suite('extract-timestamp', function (suite, test) {
 			})
 		})
 	})
+	suite('formats', function (suite, test) {
+		formatTests.forEach(function (value) {
+			suite(JSON.stringify(value.input), function (suite, test) {
+				for (const format of Object.keys(value.formats)) {
+					const expected = value.formats[format as Format]
+					test(format, function () {
+						equal(
+							stringifyTimestamp(value.input, format as Format),
+							expected,
+							'stringify result was as expected'
+						)
+					})
+				}
+			})
+		})
+	})
 	test('replace', function () {
 		equal(
 			replaceTimestamps(replaceTest.input, function (timestamp) {
 				return JSON.stringify(timestamp)
 			}),
-			replaceTest.result
+			replaceTest.replaceResult
+		)
+	})
+	test('stringify', function () {
+		equal(
+			replaceTimestamps(replaceTest.input, function (timestamp) {
+				return timestamp ? `[${stringifyTimestamp(timestamp)}]` : ''
+			}),
+			replaceTest.stringifyResult
 		)
 	})
 })