+++ HISTORY.md
@@ -1,5 +1,11 @@
 # History
 
+## v1.1.0 2020 April 7
+
+-   Updated some function signatures
+-   Handle invalid timestamps by returning `null`
+-   Added stringify formats
+
 ## v1.0.0 2020 April 6
 
 -   Initial workimg release